requirements.txt
fastecdsa==1.6.1
pysha3==1.0.2
